Stricter type checking analysis
Excluded a common rule from most sites since it would make API usage very unwieldy

Things that were actually broken:
- Notifications.onUnifiedPushMessage
  - This broke background notifications on Android!
- FuukaArchive::getPost
- Site4Chan::getBoardFlags

Add some guarantees in the form of Dio Options.responseType
Cleanup cloudflare.dart body conversions
Cleanup cloudflare.date userAgent usage

+++ lib/services/blobsync.dart
@@ -64,7 +64,7 @@ class BlobSync {
 			)
 		);
 		if (response.statusCode == 409) {
-			throw CurrentBlobOutOfDateException(path, response.data['lastRev']);
+			throw CurrentBlobOutOfDateException(path, response.data['lastRev'] as int);
 		}
 		if (response.statusCode != 200) {
 			final error = (response.data as Map?)?['error'] as String?;
@@ -75,7 +75,7 @@ class BlobSync {
 				throw BlobSyncServerException(path, 'HTTP Error ${response.statusCode}');
 			}
 		}
-		return response.data['rev'];
+		return response.data['rev'] as int;
 	}
 
 	void listenToTheirVersion({
